requirements.txt
Django==3.0.8
djangorestframework==3.11.0
psycopg2-binary==2.8.6
celery==5.0.5
redis==3.5.3
kafka-python==2.0.2